probemapper/requirements.txt
ipykernel
numpy
scipy
scikit-image
matplotlib
seaborn
h5py
nibabel
pandas
pynrrd
skan
